move place_window into winit-common
