+++ winit-common/Cargo.toml
@@ -12,6 +12,9 @@ version.workspace = true
 # Event Handler
 event-handler = []
 
+# Positioner
+positioner = ["dep:dpi"]
+
 # XKB
 wayland = ["dep:memmap2"]
 x11 = ["xkbcommon-dl?/x11", "dep:x11-dl"]
@@ -24,6 +27,7 @@ core-foundation = ["dep:block2", "dep:objc2", "dep:objc2-core-foundation"]
 foundation = ["dep:block2", "dep:objc2", "dep:objc2-foundation"]
 
 [dependencies]
+dpi = { workspace = true, optional = true }
 smol_str = { workspace = true, optional = true }
 tracing.workspace = true
 winit-core.workspace = true
