+++ winit-common/src/lib.rs
@@ -6,5 +6,7 @@ pub mod core_foundation;
 pub mod event_handler;
 #[cfg(feature = "foundation")]
 pub mod foundation;
+#[cfg(feature = "positioner")]
+pub mod positioner;
 #[cfg(feature = "xkb")]
 pub mod xkb;
